requirements.txt
Django==1.8.3
pymysql
sphinx
django-braces
django-extra-views
django-floppyforms
djangorestframework
gunicorn==19.3.0